Uploading the working version of the Defender Game
Changes: :gear:
-Enemies spawn in multiple locations! 
-Roving traders random event added. 
-Random events are random. 
-New materials to buy and sell (as well as prisoners, for when they are in the game). 
-Diplomacy is in! But it doesn't really do anything right now.
-Combat code has been refactored. Adding diplomacy required me to change combat significantly. 
-Player-score system updated.
-Minor code cleanup

Combat-specific changes: :punch:
Soldiers are now stored in an array, in which their index position denotes the direction [0] is North, [1] is East, etc.  During the combat phase, the arrays are compared and (currently) the winner is decided simply on who sent more troops to the location. Also, the player is only asked to send soldiers to locations where there will be enemies.