Big changes! :rocket:
Diplomacy now has an effect on the game. 
Melee and armour now modify combat. Melee causes the player to be more likely to be victorious, armour makes the player lose less soldiers when they lose combat. 
The player now takes prisoners when they win combat.
The winner of combat no longer kills all the opposing forces, it not kills a random amount up to the integer representing the soldiers present at a specific array index.