Nearing completion :checkered_flag:
Added archers back to combat. They modify the player's "combat power" which is used to determine a victory.
Selling and buying scraps have been removed as crafting is beyond the scope of this text-based game.
The roving trader event is now a "shady character" who is interested in buying your prisoners.
Test text removed. Some unused code was commented out and saved for future use.